// src/client/api/generated/model/storageType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Spec Studio API
 * API local của Spec Studio. Lỗi chỉ trả mã (ErrorCode); FE tự dịch.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Nơi lưu file spec
 */
export type StorageType = typeof StorageType[keyof typeof StorageType];


export const StorageType = {
  LOCAL: 'LOCAL',
  GIT: 'GIT',
  DRIVE: 'DRIVE',
} as const;